Move setupTrust to factories
