+++ src/usecases/createWard.contractTest.js
@@ -1,6 +1,6 @@
 import createWard from "./createWard";
 import AppContainer from "../containers/AppContainer";
-import setupTrust from "../testUtils/fixtures/setupTrust";
+import { setupTrust } from "../testUtils/factories";
 
 describe("createWard contract tests", () => {
   const container = AppContainer.getInstance();
