+++ src/usecases/retrieveHospitalWardVisitTotals.contractTest.js
@@ -1,5 +1,5 @@
 import AppContainer from "../containers/AppContainer";
-import setupTrust from "../testUtils/fixtures/setupTrust";
+import { setupTrust } from "../testUtils/factories";
 
 describe("retrieveHospitalWardVisitTotals contract tests", () => {
   const container = AppContainer.getInstance();
